refactor(MenuProduct) : 리뷰 반영 - 메뉴상품의 합 도메인 로직으로 이동
