+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -2,9 +2,9 @@ package kitchenpos.menus.application;
 
 import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuGroupRepository;
-import kitchenpos.menus.domain.MenuProduct;
 import kitchenpos.menus.domain.MenuRepository;
 import kitchenpos.menus.tobe.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
 import kitchenpos.products.domain.PurgomalumClient;
@@ -104,12 +104,7 @@ public class MenuService {
             .orElseThrow(NoSuchElementException::new);
         BigDecimal sum = BigDecimal.ZERO;
         for (final MenuProduct menuProduct : menu.getMenuProducts()) {
-            sum = sum.add(
-                    menuProduct.getProduct()
-                            .getPrice()
-                            .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
-                            .getValue()
-            );
+            sum = sum.add(menuProduct.getProduct().multiplyPrice(menuProduct.getQuantity()).getValue());
         }
         if (price.compareTo(sum) > 0) {
             throw new IllegalArgumentException();
@@ -124,12 +119,7 @@ public class MenuService {
             .orElseThrow(NoSuchElementException::new);
         BigDecimal sum = BigDecimal.ZERO;
         for (final MenuProduct menuProduct : menu.getMenuProducts()) {
-            sum = sum.add(
-                    menuProduct.getProduct()
-                            .getPrice()
-                            .multiply(BigDecimal.valueOf(menuProduct.getQuantity()))
-                            .getValue()
-            );
+            sum = sum.add(menuProduct.getProduct().multiplyPrice(menuProduct.getQuantity()).getValue());
         }
         if (menu.getPrice().compareTo(sum) > 0) {
             throw new IllegalStateException();
